Refactored task loading.
